% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{agent_append_user}
\alias{agent_append_user}
\title{Append a user message to an agent backend}
\usage{
agent_append_user(agent, message)
}
\arguments{
\item{agent}{An object implementing \link{AgentBackend}.}

\item{message}{User message text.}
}
\description{
Append a user message to an agent backend
}
