% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{agent_clear}
\alias{agent_clear}
\title{Clear an agent backend}
\usage{
agent_clear(agent)
}
\arguments{
\item{agent}{An object implementing \link{AgentBackend}.}
}
\description{
Clear an agent backend
}
