% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{agent_info}
\alias{agent_info}
\title{Return agent backend information}
\usage{
agent_info(agent)
}
\arguments{
\item{agent}{An object implementing \link{AgentBackend}.}
}
\description{
Return agent backend information
}
