% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\docType{data}
\name{AgentExtension}
\alias{AgentExtension}
\title{Agent extension interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
AgentExtension
}
\description{
Extensions expose a manifest plus contributed tools, commands, and hooks.
}
\keyword{datasets}
