% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bebel_append_user}
\alias{bebel_append_user}
\title{Append a ChatML user turn to a BebeLM agent transcript}
\usage{
bebel_append_user(agent, message)
}
\arguments{
\item{agent}{A \code{BebelAgent} object.}

\item{message}{User message.}
}
\value{
Invisibly returns \code{agent}.
}
\description{
Append a ChatML user turn to a BebeLM agent transcript
}
