% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bebel_async_poll}
\alias{bebel_async_poll}
\title{Poll a BebeLM async job}
\usage{
bebel_async_poll(job)
}
\arguments{
\item{job}{A \code{BebelAsyncJob}.}
}
\value{
\code{"ready"} or \code{"pending"}.
}
\description{
Poll a BebeLM async job
}
