% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_backend_clear}
\alias{bebel_backend_clear}
\title{Clear an agent backend}
\usage{
bebel_backend_clear(agent)
}
\arguments{
\item{agent}{An object implementing \code{BebelAgentBackend}.}
}
\description{
Clear an agent backend
}
