% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_backend_info}
\alias{bebel_backend_info}
\title{Return agent backend information}
\usage{
bebel_backend_info(agent)
}
\arguments{
\item{agent}{An object implementing \code{BebelAgentBackend}.}
}
\description{
Return agent backend information
}
