% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_extension_commands}
\alias{bebel_extension_commands}
\title{Return commands contributed by an extension}
\usage{
bebel_extension_commands(extension)
}
\arguments{
\item{extension}{An object implementing \code{BebelAgentExtension}.}
}
\description{
Return commands contributed by an extension
}
