% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_extension_hooks}
\alias{bebel_extension_hooks}
\title{Return hooks contributed by an extension}
\usage{
bebel_extension_hooks(extension)
}
\arguments{
\item{extension}{An object implementing \code{BebelAgentExtension}.}
}
\description{
Return hooks contributed by an extension
}
