% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_extension_manifest}
\alias{bebel_extension_manifest}
\title{Return an extension manifest}
\usage{
bebel_extension_manifest(extension)
}
\arguments{
\item{extension}{An object implementing \code{BebelAgentExtension}.}
}
\description{
Return an extension manifest
}
