% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_extension_skill_providers}
\alias{bebel_extension_skill_providers}
\title{Return skill providers contributed by an extension}
\usage{
bebel_extension_skill_providers(extension)
}
\arguments{
\item{extension}{An object implementing \code{BebelAgentExtension}.}
}
\description{
Return skill providers contributed by an extension
}
