% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_extension_tools}
\alias{bebel_extension_tools}
\title{Return tools contributed by an extension}
\usage{
bebel_extension_tools(extension)
}
\arguments{
\item{extension}{An object implementing \code{BebelAgentExtension}.}
}
\description{
Return tools contributed by an extension
}
