% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-extensions.R
\name{bebel_loop_command_catalog}
\alias{bebel_loop_command_catalog}
\title{Return a loop's command catalog}
\usage{
bebel_loop_command_catalog(loop)
}
\arguments{
\item{loop}{A \code{bebelAgentLoop}.}
}
\value{
A data frame of commands.
}
\description{
Return a loop's command catalog
}
