% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bebel_model_load}
\alias{bebel_model_load}
\title{Load a BebeLM GGUF model}
\usage{
bebel_model_load(path, num_threads = NULL)
}
\arguments{
\item{path}{Path to the GGUF weights file.}

\item{num_threads}{Optional Rayon global thread-pool size. This can only be set once per R process.}
}
\value{
A \code{BebelModel} object.
}
\description{
Load a BebeLM GGUF model
}
