% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_prompt_template_list}
\alias{bebel_prompt_template_list}
\title{List prompt templates}
\usage{
bebel_prompt_template_list(provider)
}
\arguments{
\item{provider}{A prompt-template provider.}
}
\description{
List prompt templates
}
