% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_prompt_template_render}
\alias{bebel_prompt_template_render}
\title{Render a prompt template}
\usage{
bebel_prompt_template_render(provider, name, data = list())
}
\arguments{
\item{provider}{A prompt-template provider.}

\item{name}{Template name.}

\item{data}{Template data.}
}
\description{
Render a prompt template
}
