% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{bebel_skill_load}
\alias{bebel_skill_load}
\title{Load a skill by name}
\usage{
bebel_skill_load(provider, name)
}
\arguments{
\item{provider}{A skill provider.}

\item{name}{Skill name.}
}
\description{
Load a skill by name
}
