% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\docType{data}
\name{BebelAgentBackend}
\alias{BebelAgentBackend}
\title{BebeLM agent backend interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
BebelAgentBackend
}
\description{
Backends implement the minimal transcript/generation protocol consumed by
\code{\link[=bebel_agent_loop]{bebel_agent_loop()}}.
}
\keyword{datasets}
