% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelAgentRef}
\alias{BebelAgentRef}
\title{BebeLM agent reference}
\usage{
BebelAgentRef(value = list())
}
\arguments{
\item{value}{A one-element list containing a \code{BebelAgent}.}
}
\description{
BebeLM agent reference
}
