% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelAgentRunOptions}
\alias{BebelAgentRunOptions}
\title{Agent run options}
\usage{
BebelAgentRunOptions(max_steps = integer(0), check_interrupt = logical(0))
}
\arguments{
\item{max_steps}{Maximum assistant/tool iterations.}

\item{check_interrupt}{Check for R user interrupts during synchronous generation.}
}
\description{
Agent run options
}
