% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelAsyncEventDrainOptions}
\alias{BebelAsyncEventDrainOptions}
\title{Async event drain options}
\usage{
BebelAsyncEventDrainOptions(max = NULL)
}
\arguments{
\item{max}{Optional non-negative whole-number event limit.}
}
\description{
Async event drain options
}
