% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BebelAsyncJob}
\alias{BebelAsyncJob}
\title{Background BebeLM generation job.}
\format{
An object of class \code{Rbebelm::BebelAsyncJob__bundle} (inherits from \code{savvy_Rbebelm__sealed}) of length 0.
}
\usage{
BebelAsyncJob
}
\description{
Background BebeLM generation job.
}
\keyword{datasets}
