% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelAsyncJobRef}
\alias{BebelAsyncJobRef}
\title{BebeLM async job reference}
\usage{
BebelAsyncJobRef(value = list())
}
\arguments{
\item{value}{A one-element list containing a \code{BebelAsyncJob}.}
}
\description{
BebeLM async job reference
}
