% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelAsyncWaitOptions}
\alias{BebelAsyncWaitOptions}
\title{Async wait options}
\usage{
BebelAsyncWaitOptions(
  poll_interval = integer(0),
  cancel_on_interrupt = logical(0)
)
}
\arguments{
\item{poll_interval}{Seconds to sleep between polls while a job is pending.}

\item{cancel_on_interrupt}{Whether an interrupted wait should request Rust-side
job cancellation.}
}
\description{
Async wait options
}
