% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BebelFileFinder}
\alias{BebelFileFinder}
\title{Persistent native FFF fuzzy file finder.}
\format{
An object of class \code{Rbebelm::BebelFileFinder__bundle} (inherits from \code{savvy_Rbebelm__sealed}) of length 1.
}
\usage{
BebelFileFinder
}
\description{
Persistent native FFF fuzzy file finder.
}
\keyword{datasets}
