% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BebelModel}
\alias{BebelModel}
\title{Loaded BebeLM GGUF model.}
\format{
An object of class \code{Rbebelm::BebelModel__bundle} (inherits from \code{savvy_Rbebelm__sealed}) of length 1.
}
\usage{
BebelModel
}
\description{
Loaded BebeLM GGUF model.
}
\keyword{datasets}
