% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelModelLoadOptions}
\alias{BebelModelLoadOptions}
\title{Model loading options}
\usage{
BebelModelLoadOptions(path = character(0), num_threads = NULL)
}
\arguments{
\item{path}{Path to a GGUF weights file.}

\item{num_threads}{Optional positive whole-number Rayon thread count.}
}
\description{
Model loading options
}
