% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelModelRef}
\alias{BebelModelRef}
\title{BebeLM model reference}
\usage{
BebelModelRef(value = list())
}
\arguments{
\item{value}{A one-element list containing a \code{BebelModel}.}
}
\description{
BebeLM model reference
}
