% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\docType{data}
\name{BebelPromptTemplateProvider}
\alias{BebelPromptTemplateProvider}
\title{BebeLM prompt-template provider interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
BebelPromptTemplateProvider
}
\description{
Prompt-template providers list and render named templates.
}
\keyword{datasets}
