% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelScalarText}
\alias{BebelScalarText}
\title{Scalar non-empty text}
\usage{
BebelScalarText(value = character(0))
}
\arguments{
\item{value}{A non-empty character scalar.}
}
\description{
Scalar non-empty text
}
