% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelTokenizeOptions}
\alias{BebelTokenizeOptions}
\title{Tokenizer options}
\usage{
BebelTokenizeOptions(add_bos = logical(0))
}
\arguments{
\item{add_bos}{Whether to prepend the model's beginning-of-sequence token.}
}
\description{
Tokenizer options
}
