% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BebelToolRef}
\alias{BebelToolRef}
\title{BebeLM tool reference}
\usage{
BebelToolRef(value = list())
}
\arguments{
\item{value}{A one-element list containing a \code{BebelToolSpec}.}
}
\description{
BebeLM tool reference
}
