% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{ColbertEmbeddings}
\alias{ColbertEmbeddings}
\title{One query or document's L2-normalized ColBERT token vectors.}
\usage{
ColbertEmbeddings
}
\description{
One query or document's L2-normalized ColBERT token vectors.
}
