% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ColbertEmbeddingsRef}
\alias{ColbertEmbeddingsRef}
\title{ColBERT token-vector reference}
\usage{
ColbertEmbeddingsRef(value = list())
}
\arguments{
\item{value}{A one-element list containing \code{ColbertEmbeddings}.}
}
\description{
ColBERT token-vector reference
}
