% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ColbertModelRef}
\alias{ColbertModelRef}
\title{ColBERT model reference}
\usage{
ColbertModelRef(value = list())
}
\arguments{
\item{value}{A one-element list containing a \code{ColbertModel}.}
}
\description{
ColBERT model reference
}
