% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{EmbeddingGemmaModel}
\alias{EmbeddingGemmaModel}
\title{Loaded EmbeddingGemma GGUF model.}
\format{
An object of class \code{Rbebelm::EmbeddingGemmaModel__bundle} (inherits from \code{savvy_Rbebelm__sealed}) of length 1.
}
\usage{
EmbeddingGemmaModel
}
\description{
Loaded EmbeddingGemma GGUF model.
}
\keyword{datasets}
