% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{EmbeddingGemmaModelRef}
\alias{EmbeddingGemmaModelRef}
\title{EmbeddingGemma model reference}
\usage{
EmbeddingGemmaModelRef(value = list())
}
\arguments{
\item{value}{A one-element list containing an \code{EmbeddingGemmaModel}.}
}
\description{
EmbeddingGemma model reference
}
