% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{extension_commands}
\alias{extension_commands}
\title{Return commands contributed by an extension}
\usage{
extension_commands(extension)
}
\arguments{
\item{extension}{An object implementing \link{AgentExtension}.}
}
\description{
Return commands contributed by an extension
}
