% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{extension_hooks}
\alias{extension_hooks}
\title{Return hooks contributed by an extension}
\usage{
extension_hooks(extension)
}
\arguments{
\item{extension}{An object implementing \link{AgentExtension}.}
}
\description{
Return hooks contributed by an extension
}
