% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{extension_manifest}
\alias{extension_manifest}
\title{Return an extension manifest}
\usage{
extension_manifest(extension)
}
\arguments{
\item{extension}{An object implementing \link{AgentExtension}.}
}
\description{
Return an extension manifest
}
