% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{extension_skill_providers}
\alias{extension_skill_providers}
\title{Return skill providers contributed by an extension}
\usage{
extension_skill_providers(extension)
}
\arguments{
\item{extension}{An object implementing \link{AgentExtension}.}
}
\description{
Return skill providers contributed by an extension
}
