% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{extension_tools}
\alias{extension_tools}
\title{Return tools contributed by an extension}
\usage{
extension_tools(extension)
}
\arguments{
\item{extension}{An object implementing \link{AgentExtension}.}
}
\description{
Return tools contributed by an extension
}
