% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\name{prompt_template_list}
\alias{prompt_template_list}
\title{List prompt templates}
\usage{
prompt_template_list(provider)
}
\arguments{
\item{provider}{A prompt-template provider.}
}
\description{
List prompt templates
}
