% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{rbebelm_backend_features}
\alias{rbebelm_backend_features}
\title{Return feature information reported by the loaded Rust backend.}
\usage{
rbebelm_backend_features()
}
\description{
Return feature information reported by the loaded Rust backend.
}
