% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{rbebelm_json_parse}
\alias{rbebelm_json_parse}
\title{Parse JSON into base R vectors/lists.}
\usage{
rbebelm_json_parse(text)
}
\description{
Parse JSON into base R vectors/lists.
}
\keyword{internal}
