% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{rbebelm_json_tool_result}
\alias{rbebelm_json_tool_result}
\title{Format a tool-result object as JSON.}
\usage{
rbebelm_json_tool_result(tool, ok, result = NULL, error = NULL)
}
\description{
Format a tool-result object as JSON.
}
\keyword{internal}
