% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-contracts.R
\docType{data}
\name{SkillProvider}
\alias{SkillProvider}
\title{Skill provider interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
SkillProvider
}
\description{
Skill providers list and load reusable instructions or workflow snippets.
}
\keyword{datasets}
