% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BebelAgent}
\alias{BebelAgent}
\title{Persistent BebeLM conversation agent with transcript and decode caches.}
\format{
An object of class \code{Rbebelm::BebelAgent__bundle} (inherits from \code{savvy_Rbebelm__sealed}) of length 1.
}
\usage{
BebelAgent
}
\description{
Persistent BebeLM conversation agent with transcript and decode caches.
}
\keyword{datasets}
